+++ src/assetbundles/src/js/AddTranslationsToCommerce.js
@@ -229,7 +229,7 @@
         updateSelectedEntries: function() {
             var entries = [];
 
-            $('.elements table.data tbody tr.sel[data-id]').each(function() {
+            $('.elements table.data tbody tr.sel[data-id], .elements ul.card-grid li.sel[data-id]').each(function() {
                 entries.push($(this).data('id'));
             });
 
@@ -257,17 +257,7 @@
                 return $('[name=siteId]').val();
             }
 
-            var localeMenu = $('.sitemenubtn').data('menubtn').menu;
-
-            // Figure out the initial locale
-            var $option = localeMenu.$options.filter('.sel:first');
-
-
-            if ($option.length === 0) {
-                $option = localeMenu.$options.first();
-            }
-
-            var siteId = $option.data('site-id').toString();
+            var siteId = $('.element').attr('data-site-id');
 
             return siteId;
         },