+++ src/services/fieldtranslator/MatrixFieldTranslator.php
@@ -94,10 +94,11 @@ class MatrixFieldTranslator extends GenericFieldTranslator
              * Block id changes for localised block so use $i and using same for non localised blocks merges other
              * sites non localised block to non localised block.
              */
-            $blockId = $field->getIsTranslatable() ? $i : $block->id;
+            $blockId = $field->getIsTranslatable($element) ? $i : $block->id;
             $blockData = $fieldData[$i] ?? $fieldData[$oldKey] ?? array();
 
             $post[$fieldHandle][$blockId] = array(
+                'title'             => $block->title,
                 'type'              => $block->getType()->handle,
                 'enabled'           => $block->getAttributes()['enabled'],
                 'collapsed'         => $block->getAttributes()['collapsed'],