+++ src/services/repository/OrderRepository.php
@@ -510,12 +510,13 @@ class OrderRepository
 
                     $sourceContent = json_decode($converter->xmlToJson($file->source), true);
                     $currentContent = json_decode($converter->xmlToJson($currentContent), true);
+                    if (isset($currentContent['content']) && isset($sourceContent['content'])) {
+                        $sourceContent = json_encode(array_values($sourceContent['content']));
+                        $currentContent = json_encode(array_values($currentContent['content']));
 
-                    $sourceContent = json_encode(array_values($sourceContent['content']));
-                    $currentContent = json_encode(array_values($currentContent['content']));
-
-                    if (md5($sourceContent) !== md5($currentContent)) {
-                        array_push($originalIds, $element->id);
+                        if (md5($sourceContent) !== md5($currentContent)) {
+                            array_push($originalIds, $element->id);
+                        }
                     }
                 } catch (Exception $e) {
                     throw new Exception("Source entry changes check, Error: " . $e->getMessage(), 1);